--- models.py
@@ -1,19 +1,0 @@
-from main import database
-from datetime import datetime
-
-
-class Usuario(database.Model):
-    
-    id = database.Column(database.Integer, primary_key=True)
-    username = database.Column(database.String, nullable=False)
-    email = database.Column(database.String, nullable=False, unique=True)
-    senha = database.Column(database.String, nullable=False)
-    foto_perfil = database.Column(database.String, default='default.jpg')
-    
-
-class Post(database.Model):
-    
-    id = database.Column(database.String, nullable=False)
-    titulo = database.Column(database.String, nullable=False)
-    corpo = database.Column(database.Text, nullable=False)
-    data_criação = database.Column(database.DateTime, nullable=False, default=datetime.utcnow)
