+++ mycommunity/templates/navbar.html
@@ -13,7 +13,7 @@
             <a class="nav-link active" aria-current="page" href="{{ url_for('usuarios') }}">Usuários</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="https://github.com/rychardmazarin30?tab=repositories">Contato</a>
+            <a class="nav-link active" aria-current="page" href="https://github.com/rychardmazarin30">Contato</a>
           </li>
         </ul>
         <a style="margin: 0 15px;" class="btn btn-outline-secondary" href="{{  url_for('login') }}" role="button">Login</a>
