+++ mycommunity/templates/usuarios.html
@@ -6,7 +6,7 @@
     <div class="text-white container" style="border-radius: 1rem;">
         <h1 style="text-align: center;">Usuários</h1>
         {% for user in users %}
-            <strong>{{user}}</strong>
+            <strong>{{ user.nome.capitalize() }} {{ user.sobrenome.capitalize() }}</strong>
             <hr>
         {% endfor %}
     </div>
