+++ mycommunity/templates/welcome.html
@@ -1,0 +1,12 @@
+{% extends "base.html" %}
+
+{% block body %}
+    <br>
+    <div style="text-align: center;">
+        <div class="text-white" style="border-radius: 1rem;">
+            <h1 >MY COMMUNITY</h1>
+            <h4 >Welcome, {{ usuario.nome.capitalize()}} {{ usuario.sobrenome.capitalize() }} </h4>
+        </div>
+    </div>
+
+{% endblock %}
