--- templates/contato.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-
-{% block body %}
-
-    <div class="container">
-        <h1><strong>Dúvidas: </strong>4002-8922</h1>
-    </div>
-
-{% endblock %}
